Use json measurements also if no WHERE clause is given
The error occurred, because we did not use the JSON measurments, when
a 'mvalue' query did not point to a json, that is had a json selector.
For example, mvalue.address.city.eq.Merano

The problem was, if no WHERE clause was present at all, the json
measurements where always thrown away, since no json selector was present.